Break out `test_parser.py` into test specific files
