+++ tests/parser/functions/test_length.py
@@ -1,0 +1,17 @@
+import pytest
+from tests.setup_transaction_tests import chain as s, tester as t, ethereum_utils as u, check_gas, \
+    get_contract_with_gas_estimation
+
+
+def test_test_length():
+    test_length = """
+y: bytes <= 10
+def foo(inp: bytes <= 10) -> num:
+    x = slice(inp, start=1, len=5)
+    self.y = slice(inp, start=2, len=4)
+    return len(inp) * 100 + len(x) * 10 + len(self.y)
+    """
+
+    c = get_contract_with_gas_estimation(test_length)
+    assert c.foo(b"badminton") == 954, c.foo(b"badminton")
+    print('Passed length test')
