+++ tests/parser/globals/test_globals.py
@@ -1,0 +1,18 @@
+import pytest
+from tests.setup_transaction_tests import chain as s, tester as t, ethereum_utils as u, check_gas, \
+    get_contract_with_gas_estimation
+
+def test_permanent_variables_test():
+    permanent_variables_test = """
+var: {a: num, b: num}
+def __init__(a: num, b: num):
+    self.var.a = a
+    self.var.b = b
+
+def returnMoose() -> num:
+    return self.var.a * 10 + self.var.b
+    """
+
+    c = get_contract_with_gas_estimation(permanent_variables_test, args=[5, 7])
+    assert c.returnMoose() == 57
+    print('Passed init argument and variable member test')
