+++ tests/parser/integration/test_basics.py
@@ -23,3 +23,26 @@ def foo(x: num) -> num:
     c = get_contract_with_gas_estimation(basic_code)
     assert c.foo(9) == 18
     print('Passed basic code test')
+
+
+def test_selfcall_code_3():
+    selfcall_code_3 = """
+def _hashy2(x: bytes <= 100) -> bytes32:
+    return sha3(x)
+
+def return_hash_of_cow_x_30() -> bytes32:
+    return self._hashy2("cowcowcowcowcowcowcowcowcowcowcowcowcowcowcowcowcowcowcowcowcowcowcowcowcowcowcowcowcowcow")
+
+def _len(x: bytes <= 100) -> num:
+    return len(x)
+
+def returnten() -> num:
+    return self._len("badminton!")
+    """
+
+    c = get_contract_with_gas_estimation(selfcall_code_3)
+    assert c.return_hash_of_cow_x_30() == u.sha3(b'cow' * 30)
+    assert c.returnten() == 10
+
+    print("Passed single variable-size argument self-call test")
+
